requirements.txt
cassandra-driver==3.10
cql==1.4.0
thrift==0.10.0
Flask==0.10.1
Flask-Classy==0.6.10
Flask-Cors==2.1.0
Werkzeug==0.10.4
tornado==4.1
